# Mimir query engine: improve performance for case where right side of binary operation has fewer series than the left side