--- a/pkg/streamingpromql/testdata/ours/binary_operators.test
+++ b/pkg/streamingpromql/testdata/ours/binary_operators.test
@@ -142,3 +142,17 @@ eval_fail range from 0 to 42m step 6m left_side * on (env) right_side
   # TODO: expected_message found duplicate series for the match group {env="test"} on the right hand-side of the operation: [{__name__="right_side", env="test", pod="b"}, {__name__="right_side", env="test", pod="a"}];many-to-many matching not allowed: matching labels must be unique on one side
 
 clear
+
+# One-to-one matching with more series on left side than right (and vice-versa)
+# We have an optimsation that favours the smaller side, these tests ensure it behaves correctly.
+load 1m
+  single_series{env="prod"} 100 200 300
+  many_series{env="canary"} 10 20 30
+  many_series{env="prod"} 40 50 60
+  many_series{env="test"} 70 80 90
+
+eval range from 0 to 2m step 1m single_series - many_series
+  {env="prod"} 60 150 240
+
+eval range from 0 to 2m step 1m many_series - single_series
+  {env="prod"} -60 -150 -240