#### What this PR does

This PR improves the performance of binary operations for queries where the right side has fewer series than the left side.

Benchmark results for a query where the left side (`a_2000`) has 2000 series, and the right side has one series:

```
                                                                                    │ original.txt  │             after.txt              │
                                                                                    │    sec/op     │   sec/op     vs base               │
Query/a_2000_-_b_2000{l="1234"},_instant_query/streaming-10                            6.917m ±  4%   6.327m ± 1%   -8.53% (p=0.002 n=6)
Query/a_2000_-_b_2000{l="1234"},_range_query_with_100_steps/streaming-10              10.391m ±  6%   9.670m ± 1%   -6.94% (p=0.002 n=6)
Query/a_2000_-_b_2000{l="1234"},_range_query_with_1000_steps/streaming-10              47.98m ±  4%   46.55m ± 1%   -2.98% (p=0.041 n=6)
```

Latency and peak memory utilisation is unchanged for other benchmarks, and peak memory utilisation is only marginally affected for the benchmark above.

#### Which issue(s) this PR fixes or relates to

Resolves a [TODO](https://github.com/grafana/mimir/pull/8096/files#diff-ccf54e809ccc9d8e8a83091d9fd45f64ba2e3fadb87a02dae00ceb431b62fd4dR159-R160) from https://github.com/grafana/mimir/pull/8096

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [n/a] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.